+++ docs/CHANGELOG.md
@@ -13,6 +13,11 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Added
+- The `cognite.pygen.utils.MockGenerator` now generate properties for edges that have properties.
+
+
 ## [0.99.40] - 24-10-26
 ### Added
 - Support for using `NodeId` and `tuple[str, str]` as well as sequences of these in the `.retrieve`